% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_group_averages_matrix.R
\name{calculate_group_averages_matrix}
\alias{calculate_group_averages_matrix}
\title{Group Averages Matrix}
\usage{
calculate_group_averages_matrix(
  raw_data,
  categorical_columns,
  weights_column,
  debt_column
)
}
\arguments{
\item{raw_data}{a data set of credibility.}

\item{categorical_columns}{categorical column of data set.}

\item{weights_column}{weights column of data set.}

\item{debt_column}{credit dept column of data set.}
}
\value{
group averages matrix
}
\description{
Group Averages Matrix
}
\examples{
raw_data <- debt

categorical_columns <- c(1,2)

weights_column <- 3

debt_column <- 4

calculate_group_averages_matrix(raw_data, categorical_columns, weights_column, debt_column)
}
