% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_obs_and_group_weights.R
\name{calculate_obs_and_group_weights}
\alias{calculate_obs_and_group_weights}
\title{Repeats of observations}
\usage{
calculate_obs_and_group_weights(
  raw_data,
  categorical_columns,
  weights_column,
  debt_column
)
}
\arguments{
\item{raw_data}{a data set of credibility.}

\item{categorical_columns}{categorical column of data set.}

\item{weights_column}{weights column of data set.}

\item{debt_column}{credit dept column of data set.}
}
\value{
This function returns categorical group sizes.
}
\description{
Repeats of observations
}
\examples{
raw_data <- debt

categorical_columns <- c(1,2)

weights_column <- 3

debt_column <- 4

calculate_obs_and_group_weights(raw_data, categorical_columns, weights_column, debt_column)
}
