% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_variance_and_std.R
\name{calculate_variance_and_std}
\alias{calculate_variance_and_std}
\title{Variance and Standard Deviation}
\usage{
calculate_variance_and_std(
  raw_data,
  categorical_columns,
  weights_column,
  debt_column
)
}
\arguments{
\item{raw_data}{a data set of credibility.}

\item{categorical_columns}{categorical column of data set.}

\item{weights_column}{weights column of data set.}

\item{debt_column}{credit dept column of data set.}
}
\value{
variance and sd.
}
\description{
Variance and Standard Deviation
}
\examples{
raw_data <- debt

categorical_columns <- c(1,2)

weights_column <- 3

debt_column <- 4

calculate_variance_and_std(raw_data, categorical_columns, weights_column, debt_column)
}
