% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_diff_matrix_with_vector.R
\name{col_diff_matrix_with_vector}
\alias{col_diff_matrix_with_vector}
\title{Column Wise Matrix Diff}
\usage{
col_diff_matrix_with_vector(m, vec)
}
\arguments{
\item{m}{is a matrix}

\item{vec}{is a vector}
}
\value{
This function returns a \code{num} matrix.
}
\description{
This function returns of the column wise difference between the m matrix and the vector v
}
