% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_data.R
\name{control_data}
\alias{control_data}
\title{Data checker}
\usage{
control_data(x)
}
\arguments{
\item{x}{a dataset.}
}
\value{
This function checks whether \code{dataset} has at least 2 features in categorical format.
}
\description{
Throws an error message if at least 2 features is not in categorical format.
}
