% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debt.R
\docType{data}
\name{debt}
\alias{debt}
\title{Debt Data}
\format{
A data frame of 106 rows and 4 columns
\describe{
 \item{bank}{categorical data of bank type. Bank type includes four subcategory such as State Banks, Deposit Banks, Foreign Banks and  Privately Owned Deposit Banks}
 \item{loan}{categorical data of dept type. Loan type includes three subcategory such as non-performing vehicle, home, and consumer loan.}
 \item{weights}{Numeric values of weights}
 \item{debt}{Numeric values of debt}
}
}
\usage{
debt
}
\description{
A real data which published by Turkey Banking Regulation and Supervisory Board <https://www.bddk.org.tr/BultenAylik/en>.
}
\keyword{datasets}
