% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_matrix_rows_with_vector.R
\name{div_matrix_rows_with_vector}
\alias{div_matrix_rows_with_vector}
\title{Row Wise Matrix Division}
\usage{
div_matrix_rows_with_vector(m, vec)
}
\arguments{
\item{m}{is a matrix}

\item{vec}{is a vector}
}
\value{
This function returns a \code{num} matrix.
}
\description{
This function returns of the row wise division of the m matrix and the vector v.
}
