% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult_matrix_cols_with_vector.R
\name{mult_matrix_cols_with_vector}
\alias{mult_matrix_cols_with_vector}
\title{Column Wise Matrix Multiplication}
\usage{
mult_matrix_cols_with_vector(m, vec)
}
\arguments{
\item{m}{is a matrix}

\item{vec}{is a vector}
}
\value{
This function returns a \code{num} matrix.
}
\description{
This function returns of the column wise multiplication of the m matrix and the vector v.
}
