% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_diff_matrix_with_vector.R
\name{row_diff_matrix_with_vector}
\alias{row_diff_matrix_with_vector}
\title{Row Wise Matrix Diff}
\usage{
row_diff_matrix_with_vector(m, vec)
}
\arguments{
\item{m}{is a matrix}

\item{vec}{is a vector}
}
\value{
This function returns a \code{num} matrix.
}
\description{
This function returns of the row wise difference between the m matrix and the vector v
}
