% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_names.R
\name{save_names}
\alias{save_names}
\title{Get names}
\usage{
save_names(raw_data, categorical_columns)
}
\arguments{
\item{raw_data}{a data set of credibility.}

\item{categorical_columns}{categorical column of data set.}
}
\value{
returns categorical variables' unique values and column names of data set.
}
\description{
Get names
}
\examples{
raw_data <- debt

categorical_columns <- c(1,2)

save_names(raw_data, categorical_columns)
}
